Added a uuid generator, so now you can start up a bunch of armstrong instances and they will all receive messages in turn! I added a helper function that lets you return text regardless of the message that your actor receives. You can still catch your messages like normal but you will need to send them off to Actor[:replier] and that is not as clean.
Consequently, armstrong_test.rb is now SMALLER.
It can actually be 2 lines (including the require)
but that would be irresponsible for demonstrating
a succinct example.

Still need to be able to wrap this thing nicely
so people don't have to go around writing Procs
all over the place. And dynamic urls. That needs
to happen.