Parameters in paths are now allowed!
Be careful, there is some really flakey actor
behavior. I think there's a race condition
somewhere or that mongrel2 is sending me garbage,
but right now, it might do this thing where the 
actors might not get registered and the messages
will be sent into the void or worse yet, 
@conn.receive will just keep getting nil data.
It only gets nil data when ZMQ gives it an empty
string message. So something wonky is going on.
Be careful